% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_test.R
\name{adf_test}
\alias{adf_test}
\title{ADF Test for Random Walk}
\usage{
adf_test(X)
}
\arguments{
\item{X}{Numeric vector, the series of log returns or prices.}
}
\value{
A list containing the ADF statistic, p-value, and conclusion of the test.
}
\description{
Performs the Augmented Dickey-Fuller test on a given series to test for a unit root.
}
\examples{
prices <- c(100, 101, 102, 103, 102, 101)
adf_test(prices)
}
