% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTest.R
\name{runs_test}
\alias{runs_test}
\title{Runs Test for Random Walk}
\usage{
runs_test(X)
}
\arguments{
\item{X}{Numeric vector, the series of log returns or prices.}
}
\value{
A list containing the z-score and p-value of the runs test.
}
\description{
Performs the runs test on a given series to test the random walk hypothesis.
}
\examples{
returns <- c(100, 101, 102, 103, 102, 101)
runs_test(returns)
}
