% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varRatioTest.R
\name{variance_ratio_test}
\alias{variance_ratio_test}
\title{Variance Ratio Test}
\usage{
variance_ratio_test(S, q, types)
}
\arguments{
\item{S}{Numeric vector, the prices series.}

\item{q}{Integer, the lag number.}

\item{types}{String, choose from \link{homo, hetero}}
}
\value{
A list containing the z-score and p-value of the variance ratio test.
}
\description{
Performs the variance ratio test on a given series of log prices to test the random walk hypothesis.
}
\examples{
log_prices <- log(c(100, 101, 102, 103, 102, 101))
variance_ratio_test(prices, 2)
}
